+++ api/src/engine/engine.service.ts
@@ -2,14 +2,18 @@ import { Injectable, Logger } from '@nestjs/common';
 import { StepRegistry } from '../registry/stepRegistry';
 import { Engine } from './engine';
 import { IContext, IFlow, IStepConstructor } from '../types';
+import { SecretsService } from '../services/secrets.service';
 
 @Injectable()
 export class EngineService {
   private readonly logger = new Logger(EngineService.name);
   private readonly engine: Engine;
 
-  constructor(private readonly registry: StepRegistry) {
-    this.engine = new Engine(this.registry);
+  constructor(
+    private readonly registry: StepRegistry,
+    private readonly secretsService: SecretsService,
+  ) {
+    this.engine = new Engine(this.registry, this.secretsService);
   }
 
   async runFlow(flow: IFlow, vars: Record<string, any>): Promise<IContext> {