+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.0.8
 - Implement send_email step with Gmail (#38)
 - Add a section for flows that are running in the background (#40)
+- Move the Options button to a different location and include the Streamly logo (#41)
 
 ## 0.0.7